Templated Output
This patch adds templated output to the REX-Ray CLI. Using the existing
"--format" flag (where available), the value "tmpl" will use one of the
default templates for a given object type, such as a volume or array of
volumes. However, if the value of "--format" is neither "json", "yaml",
or "tmpl" then it will be treated as a custom Golang template to be used
to print the output. Read more about Go templates at
https://golang.org/pkg/text/template/.

A few additional notes:

  * All templated output is pushed through a Go tabwriter
    (https://golang.org/pkg/text/tabwriter/) in order to format output
    in space-delimited columns as tabular data.

  * The Hugo sort function (https://gohugo.io/templates/functions/) is
    available to custom templates specified by the "--format" flag. The
    default templates for volume and snapshot lists sort elements by
    the "Name" field.